Add Dockerfile for postgres
