+++ MansionLearning/package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "pg-promise": "^8.4.4",
     "puppeteer": "^1.3.0"
   }
 }
